explicit cast uint32 to uint16; add more precise document on `rank_many(begin,end,ans)`

+++ src/containers/run.c
@@ -884,13 +884,13 @@ int run_container_rank(const run_container_t *container, uint16_t x) {
     return sum;
 }
 uint32_t run_container_rank_many(const run_container_t *container, uint64_t start_rank, const uint32_t* begin, const uint32_t* end, uint64_t* ans){
-    const uint16_t high = (*begin) >> 16;
+    const uint16_t high = (uint16_t)((*begin) >> 16);
     const uint32_t* iter = begin;
     int sum = 0;
     int i = 0;
     for(;iter != end; iter++) {
         uint32_t x = *iter;
-        uint16_t xhigh = x >> 16;
+        uint16_t xhigh = (uint16_t)(x >> 16);
         if(xhigh != high) return iter - begin; // stop at next container
 
         uint32_t x32 = x & 0xFFFF;
